set arate to int #371

+++ veejay-current/veejay-server/libveejay/vj-lib.h
@@ -103,7 +103,7 @@ typedef struct
 	int height;
 	float fps;
 	int active;
-	long arate;
+	int arate;
 	int  achans;
 	int  abits;
 	int  abps;
